Standalone channel writter
